Antes de mudar para 6 provas por tipo

+++ src/main/java/br/com/fireware/bpchoque/model/def/TesteFisico.java
@@ -60,7 +60,7 @@ public class TesteFisico {
 	 * ="pessoaId")) private List<PessoaDef> pessoas;
 	 */
 
-	private BigDecimal notaAprovacao;
+	private Double notaAprovacao;
 
 	@PrePersist
 	@PreUpdate
